fix from_json, extract routes to files
